LegoClassifierUI/requirements.txt
Pillow>=10.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
pyserial>=3.5

crcmod~=1.7
opencv-python